pyproject.toml
[tool.poetry]
name = "jup-ag-sdk"
version = "0.0.3"
description = "Python SDK for Jupiter Exchange APIs."
authors = ["Fiji <charismoutafidis@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/cmoutafidis/jup-ag-sdk"
repository = "https://github.com/cmoutafidis/jup-ag-sdk"
keywords = ["solana", "jupiter", "sdk"]
packages = [{include = "jup_ag_sdk"}]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[tool.poetry.dependencies]
python = "^3.9"
httpx = "^0.28.1"
pydantic = "^2.11.3"
solders = "^0.26.0"
base58 = "^2.1.1"
solana = "^0.36.6"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
responses = "^0.25.6"
flake8 = "^7.1.1"
black = "^25.1.0"
isort = "^6.0.0"
mypy = "^1.15.0"
pre-commit = "^4.1.0"
pytest-cov = "^6.1.1"
bandit = "^1.7"
safety = "^2.3"
python-dotenv = "^1.1.0"

[tool.black]
line-length = 88
target-version = ['py39', 'py310']

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 88
exclude = ["__pycache__", "build/", "dist/"]

[tool.mypy]
strict = true
disallow_untyped_defs = true
disable_error_code = ["annotation-unchecked"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

